% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{proof_authenticate}
\alias{proof_authenticate}
\title{Authenticate with PROOF API}
\usage{
proof_authenticate(username, password)
}
\arguments{
\item{username}{(character) HutchNet username}

\item{password}{(character) HutchNet password}
}
\value{
A single token (character) for bearer authentication with
the PROOF API
}
\description{
Authenticate with PROOF API
}
\examples{
# Sys.getenv("PROOF_TOKEN")
# x <- proof_authenticate("username", "password")
# Sys.getenv("PROOF_TOKEN")
}
