% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{proof_info}
\alias{proof_info}
\title{Get information about PROOF server}
\usage{
proof_info()
}
\value{
A list with fields:
\itemize{
\item \code{branch} (character): git branch of API
\item \code{commit_sha} (character): SHA of the git commit of the API
\item \code{short_commit_sha} (character): the first eight characters of \code{commit_sha}
\item \code{commit_message} (character): the commit message of the API's most recent commit
\item \code{tag} (character): tag of most recent commit; if this does not contain a hyphen, it's a release version
}
}
\description{
Get information about PROOF server
}
