% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel.R
\name{proof_job_cancel}
\alias{proof_job_cancel}
\title{Delete PROOF Cromwell server}
\usage{
proof_job_cancel()
}
\value{
A list with fields \code{job_id} and \code{info}
}
\description{
Delete PROOF Cromwell server
}
\details{
Does not return PROOF/Cromwell server URL, for that you have to
periodically call \code{\link[=proof_job_status]{proof_job_status()}}, or wait for the email from the
PROOF API
}
\references{
\url{https://github.com/FredHutch/proof-api#delete-cromwell-server}
}
