% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_status.R
\name{proof_job_status}
\alias{proof_job_status}
\title{Get PROOF API job status - is job running, what's its URL...}
\usage{
proof_job_status()
}
\value{
A list with slots:
\itemize{
\item \code{canJobStart}
\item \code{jobStatus}
\item \code{cromwellUrl}
}
}
\description{
Get PROOF API job status - is job running, what's its URL...
}
\references{
\url{https://github.com/FredHutch/proof-api#get-cromwell-server}
}
