% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeout.R
\name{proof_timeout}
\alias{proof_timeout}
\title{Set the timeout for all requests to the PROOF API}
\usage{
proof_timeout(sec = 10)
}
\arguments{
\item{sec}{(integer/numeric) number of seconds after which
requests will timeout. default: 10 sec (10000 ms)}
}
\value{
nothing, side effect of setting the timeout for requests
}
\description{
Set the timeout for all requests to the PROOF API
}
\references{
\url{https://httr.r-lib.org/reference/timeout.html}
}
