% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proofr-package.R
\docType{package}
\name{proofr-package}
\alias{proofr}
\alias{proofr-package}
\title{proofr: Client for the PROOF API}
\description{
Client for the PROOF API.
}
\author{
\strong{Maintainer}: Scott Chamberlain \email{sachamber@fredhutch.org} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID})

}
\keyword{internal}
