% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{proof_authenticate}
\alias{proof_authenticate}
\title{Authenticate with PROOF API}
\usage{
proof_authenticate(username, password)
}
\arguments{
\item{username}{(character) HutchNet username}

\item{password}{(character) HutchNet password}
}
\value{
A single token (character) for bearer authentication with
the PROOF API
}
\description{
Authenticate with PROOF API
}
\details{
We strongly recommend to not supply \code{password} as plain text like
\code{proof_authenticate(username = "jane", password = "mypassword")}, but rather
pull in your password from an environment variable stored outside of R like
\code{proof_authenticate(username = "jane", password = Sys.getenv("HUTCH_PWD"))}
}
\section{Timeout}{

If the PROOF API is unavailable, this function will timeout after
20 seconds. Contact the package maintainer if you get a timeout error.
See \code{\link[=proof_timeout]{proof_timeout()}}.
}

