% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel.R
\name{proof_cancel}
\alias{proof_cancel}
\title{Cancel/stop PROOF Cromwell server}
\usage{
proof_cancel()
}
\value{
On success, a list with a single field:
\itemize{
\item \code{message} (character)
}

If run when there's no Cromwell server running, an HTTP error
}
\description{
Cancel/stop PROOF Cromwell server
}
\section{Timeout}{

If the PROOF API is unavailable, this function will timeout after
5 seconds. Contact the package maintainer if you get a timeout error.
See \code{\link[=proof_timeout]{proof_timeout()}}.
}

\references{
\url{https://github.com/FredHutch/proof-api#delete-cromwell-server}
}
