% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel.R
\name{proof_cancel}
\alias{proof_cancel}
\title{Cancel/stop PROOF Cromwell server}
\usage{
proof_cancel()
}
\value{
A list with a single field:
\itemize{
\item \code{message} (character)
}
}
\description{
Cancel/stop PROOF Cromwell server
}
\references{
\url{https://github.com/FredHutch/proof-api#delete-cromwell-server}
}
