% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{proof_header}
\alias{proof_header}
\title{Get header for PROOF API calls}
\usage{
proof_header(req, token = NULL)
}
\arguments{
\item{req}{An \code{httr2} request. required}

\item{token}{(character) PROOF API token. optional}
}
\value{
An \code{httr2_request} S3 class adding an HTTP header for
\code{Authorization} with the value in \code{token}
}
\description{
Get header for PROOF API calls
}
