% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{proof_header}
\alias{proof_header}
\title{Get header for PROOF API calls}
\usage{
proof_header(token = NULL)
}
\arguments{
\item{token}{(character) PROOF API token. optional}
}
\value{
A \code{request} S3 class with the HTTP header that can be passed
to \code{httr::GET()}, \code{httr::POST()}, etc.
}
\description{
Get header for PROOF API calls
}
