% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{proof_start}
\alias{proof_start}
\title{Start PROOF Cromwell server}
\usage{
proof_start(slurm_account = NULL)
}
\arguments{
\item{slurm_account}{(character) PI name in the form \code{last_f}, where last
is the last name of the PI and f is the first name, e.g., "Jane Doe"
would be \code{doe_j}; only needed if user is in more than one SLURM account.
optional}
}
\value{
A list with fields:
\itemize{
\item \code{job_id} (character) - the job ID
\item \code{info} (character) - message
}
}
\description{
Start PROOF Cromwell server
}
\details{
Does not return PROOF/Cromwell server URL, for that you have to
periodically call \code{\link[=proof_status]{proof_status()}}, or wait for the email from the
PROOF API. See the link to proof-api for more details about the
slurm account.
}
\section{Timeout}{

If the PROOF API is unavailable, this function will timeout after
5 seconds. Contact the package maintainer if you get a timeout error.
See \code{\link[=proof_timeout]{proof_timeout()}}.
}

\references{
\url{https://github.com/FredHutch/proof-api/#post-cromwell-server}
}
