% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{proof_start}
\alias{proof_start}
\title{Start PROOF Cromwell server}
\usage{
proof_start(pi_name = NULL)
}
\arguments{
\item{pi_name}{(character) PI name in the form last_f; only needed if user
is in more than one SLURM account}
}
\value{
A list with fields:
\itemize{
\item \code{job_id} (character)
\item \code{info} (character)
}
}
\description{
Start PROOF Cromwell server
}
\details{
Does not return PROOF/Cromwell server URL, for that you have to
periodically call \code{\link[=proof_status]{proof_status()}}, or wait for the email from the
PROOF API
}
\references{
\url{https://github.com/FredHutch/proof-api#post-cromwell-server}
}
