% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{proof_start}
\alias{proof_start}
\title{Start PROOF Cromwell server}
\usage{
proof_start(slurm_account = NULL, token = NULL)
}
\arguments{
\item{slurm_account}{(character) PI name in the form \code{last_f}, where last
is the last name of the PI and f is the first name, e.g., "Jane Doe"
would be \code{doe_j}; only needed if user is in more than one SLURM account.
optional}

\item{token}{(character) token for bearer authentication with
the PROOF API. default is \code{NULL}. we do not recommend passing your token
here as a string. Either pass it using \code{\link[=Sys.getenv]{Sys.getenv()}} or save your
token as an env var and then passing nothing to this param and we'll find
it}
}
\value{
A list with fields:
\itemize{
\item \code{job_id} (character) - the job ID
\item \code{info} (character) - message
}
}
\description{
Start PROOF Cromwell server
}
\details{
Does not return PROOF/Cromwell server URL, for that you have to
periodically call \code{\link[=proof_status]{proof_status()}}, or wait for the email from the
PROOF API
}
\section{Cromwell Server uptime}{

The Cromwell server started by this function will run for 7 days
and then shut down. However, your data persists across different
instances of Cromwell servers created by running this function.
You do however need to have a Cromwell server up and running to
retrieve data.
}

\section{Timeout}{

If the PROOF API is unavailable, this function will timeout after
20 seconds. Contact the package maintainer if you get a timeout error.
See \code{\link[=proof_timeout]{proof_timeout()}}.
}

