% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{proof_status}
\alias{proof_status}
\title{Get PROOF API job status - is job running, what's its URL...}
\usage{
proof_status(wait = FALSE, token = NULL)
}
\arguments{
\item{wait}{(logical) if \code{TRUE} wait for the server to be ready to
interact with. if \code{FALSE} return immediately, then you'll want to call
this function again until you get the server URL}

\item{token}{(character) token for bearer authentication with
the PROOF API. default is \code{NULL}. we do not recommend passing your token
here as a string. Either pass it using \code{\link[=Sys.getenv]{Sys.getenv()}} or save your
token as an env var and then passing nothing to this param and we'll find
it}
}
\value{
A list with fields:
\itemize{
\item \code{canJobStart} (logical): can a job to make a Cromwell server be started?
\code{FALSE} if server already running; \code{TRUE} if server not running
\item \code{jobStatus} (character): status of the job that controls the Cromwell
server. e.g., "RUNNING". \code{NULL} if server not running
\item \code{cromwellUrl} (character): the URL for the Cromwell server. \code{NULL} if
server not running
\item \code{jobInfo} (list): metadata on the Cromwell server. All slots \code{NULL}
if server not running
}
}
\description{
Get PROOF API job status - is job running, what's its URL...
}
\section{Timeout}{

If the PROOF API is unavailable, this function will timeout after
20 seconds. Contact the package maintainer if you get a timeout error.
See \code{\link[=proof_timeout]{proof_timeout()}}.
}

