% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proofr-package.R
\docType{package}
\name{proofr-package}
\alias{proofr}
\alias{proofr-package}
\title{proofr: Client for the PROOF API}
\description{
Client for the PROOF API.
}
\section{Base URL for PROOF API}{

The base URL for the PROOF API can be changed by setting the environment
variable \code{PROOF_API_BASE_URL}. It can be set for an R session or for
function by function use as we check that env var in each function call
to the API
}

\seealso{
Useful links:
\itemize{
  \item \url{http://getwilds.org/proofr/}
  \item \url{https://github.com/getwilds/proofr}
}

}
\author{
\strong{Maintainer}: Scott Chamberlain \email{sachamber@fredhutch.org} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID})

}
\keyword{internal}
